Add $clean_source = 0; in sbuildrc
We don't want to clean the build tree before building, because
that requires some build-dependencies which we may not have
installed.

The $clean_source option is present only in newer versions of
sbuild, so we also need to install the latest jessie-backports
version os sbuild, which this patch also does.

Also fix unix rights for ${BUILDER_HOMEDIR}/.gnupg before
generating keys, and install haveged to make sure we have enough
entropy to generate GPG keys and such.

Change-Id: Ie478e97ea0a5a9a908ecc915bf1333de87fb2ce1